Use a JDK 8 toolchain when compiling Groovy
Groovy does not support the release flag. In order to make sure we are compiling against
the Java 8 standard library when compiling groovy code that targets JVM 8 bytecode, we must
either set the bootstrap classpath to a java 8 stdlib or use a java 8 toolchain. A toolchain
is easier to set up.

The issue was that before when we used JDK 21 to compile to source/target 8, we still compiled
against 21 standard library. When running on JVM 8, there was some evidence left in the bytecode
of the java 21 standard lib, for example some of our test fixtures extend the Thread class, and
its compiled bytecode referenced classes that only existed after Java 8.

The concequence is that we need two JDKs to build gradle. Java 8 for groovy code, and Java 21
the daemon and java compilers, since we can use the release flag for java.